chore(php): Rename zend components
