+++ php/laminas/config.yaml
@@ -1,0 +1,7 @@
+framework:
+  website: getlaminas.org
+  version: 3.1
+  
+before_command:
+  - mkdir /usr/src/app/data
+  - chmod 777 -R /usr/src/app/data
