+++ php/laminas/global.php
