+++ php/laminas/modules.config.php
@@ -11,7 +11,8 @@
  * This should be an array of module namespaces used in the application.
  */
 return [
-    'Zend\Router',
-    'Zend\Validator',
+    'Laminas\ZendFrameworkBridge',
+    'Laminas\Router',
+    'Laminas\Validator',
     'Application',
 ];
