+++ php/mezzio/Home.php
@@ -6,9 +6,9 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Zend\Diactoros\Response\TextResponse;
+use Laminas\Diactoros\Response\TextResponse;
 
-class User implements MiddlewareInterface
+class Home implements MiddlewareInterface
 {
     /**
      * Process an incoming server request.
