--- php/zend-expressive/composer.json
@@ -1,17 +1,0 @@
-{
-  "require": {
-    "zendframework/zend-component-installer": "*",
-    "zendframework/zend-config-aggregator": "*",
-    "zendframework/zend-diactoros": "*",
-    "zendframework/zend-expressive": "~3.2.0",
-    "zendframework/zend-expressive-helpers": "*",
-    "zendframework/zend-stdlib": "*",
-    "zendframework/zend-servicemanager": "*",
-    "zendframework/zend-expressive-fastroute": "*"
-  },
-  "autoload": {
-    "psr-4": {
-      "App\\": "src/App/src/"
-    }
-  }
-}
