--- php/zend-framework/composer.json
@@ -1,10 +1,0 @@
-{
-  "require": {
-    "zendframework/zend-mvc": "~3.1.0"
-  },
-  "autoload": {
-    "psr-4": {
-      "Application\\": "module/Application/src/"
-    }
-  }
-}
